requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.36
psycopg[binary]>=3.1.0       # Python 3.14 compatible; replaces asyncpg
alembic==1.13.1

# Config & validation
pydantic-settings==2.3.0
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0          # used directly; passlib dropped (incompatible with bcrypt 5.x)

# AI / LLM
openai>=1.35.0
anthropic>=0.28.0

# Vector DB
qdrant-client>=1.9.1

# Object storage
boto3>=1.34.0

# Document parsing
pymupdf>=1.24.0
python-docx>=1.1.0
unstructured>=0.14.0
# unstructured extras (install separately if needed): pip install "unstructured[pdf,docx]"

# Reranking
sentence-transformers>=3.0.0

# Utilities
tiktoken>=0.7.0
aiofiles>=23.2.1
httpx>=0.27.0
numpy>=1.26.0
